+++ snapshot-tests-core/src/main/java/de/skuzzle/test/snapshots/EnableSnapshotTests.java
@@ -64,8 +64,8 @@ import de.skuzzle.test.snapshots.junit5.JUnit5SnapshotExtension;
  * <h3>Parameterized tests</h3>
  * <p>
  * Snapshot tests can be combined with JUnit5's parameterized tests, but only when you
- * provide an explict name for each snapshot assertion. With the default automatic
- * snapshot naming scheme, snapshots would otherwise be overriden for each parameterized
+ * provide an explicit name for each snapshot assertion. With the default automatic
+ * snapshot naming scheme, snapshots would otherwise be overridden for each parameterized
  * execution.
  *
  * <pre>
@@ -84,9 +84,26 @@ import de.skuzzle.test.snapshots.junit5.JUnit5SnapshotExtension;
  *                 .assertThat(actual).as...;
  * </pre>
  *
+ * <h3>Updating snapshots</h3>
+ * <p>
+ * Snapshots can become outdated when your code under test changes on purpose. In that
+ * case you can advice the framework to override existing snapshots with your code under
+ * test's actual result by placing the annotation {@link ForceUpdateSnapshots} on either
+ * the whole snapshot test class or on a single test method.
+ *
+ * <h3>Orphaned snapshots</h3>
+ * <p>
+ * Snapshot files can become orphans if, for example you rename a test class/method or you
+ * change the snapshot assertions within a test. This framework comes with a sophisticated
+ * approach for detecting those orphaned files. By default, we will log a warning with the
+ * found orphan. You can temporarily place the {@link DeleteOrphans} annotation on a
+ * snapshot test class to have those files deleted automatically.
+ *
  * @author Simon Taddiken
  * @see Snapshot
  * @see SnapshotNaming
+ * @see DeleteOrphans
+ * @see ForceUpdateSnapshots
  */
 @Retention(RUNTIME)
 @Target({ TYPE, METHOD })