+++ snapshot-tests-core/src/main/java/de/skuzzle/test/snapshots/ForceUpdateSnapshots.java
@@ -10,8 +10,10 @@ import java.lang.annotation.Target;
 import org.apiguardian.api.API;
 import org.apiguardian.api.API.Status;
 
+import de.skuzzle.test.snapshots.SnapshotDsl.ChooseAssertions;
+
 /**
- * Can be temporarily placed on a test class or test method to forcefully update
+ * Can be <em>temporarily</em> placed on a test class or test method to forcefully update
  * snapshots.
  * <p>
  * When placed on a test class, snapshots in all tests within that class will be
@@ -21,9 +23,13 @@ import org.apiguardian.api.API.Status;
  * <p>
  * This annotation replaces the {@link EnableSnapshotTests#forceUpdateSnapshots()} flag
  * because it is easier to use.
+ * <p>
+ * Besides using this annotations, snapshots can also be updated globally by passing the
+ * system property <code>forceUpdateSnapshots</code> (case insensitive) to the JVM.
  *
  * @deprecated This annotation is <b>NOT</b> deprecated. Deprecation serves only to mark
  *             this annotation in your IDE as it should only be used temporarily.
+ * @see ChooseAssertions#justUpdateSnapshot()
  * @author Simon Taddiken
  * @since 1.1.0
  * @apiNote <b>Warning:</b> While this is annotation is present, all tests containing